Comment: 상품, 게시글 등록 전체적 코드/파일 정리(#277)
