+++ src/components/PostProductWriting/PostWriting.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
+import * as T from "./writingUI.styled";
 
-import * as T from "./commonCss.styled";
 import Textarea from '../common/Textarea';
 import Button from '../common/Button/Button';
 
-export default function Posting({ handleInputChange, description, handleError }) {
+export default function PostWriting({ handleInputChange, description, handleError }) {
   return (
     <T.ContentUploadWrap>
       <T.InputWrap>
@@ -24,7 +24,7 @@ export default function Posting({ handleInputChange, description, handleError })
       <Button
         type="submit"
         height="56px"
-        text={"업로드 하기"}
+        text={"게시글 업로드 하기"}
         br="4px"
         onClick={handleError}
       />
