+++ src/components/common/imageOptimization.js
@@ -1,6 +1,6 @@
 import UploadImage from "../../api/UploadImage";
 
-export const handleDataForm = async (dataURI, name) => {
+export const handleDataForm = async (dataURI, name, setImageWrap, setLoading) => {
   const byteString = atob(dataURI.split(",")[1]);
   const ab = new ArrayBuffer(byteString.length);
   const ia = new Uint8Array(ab);
@@ -12,4 +12,11 @@ export const handleDataForm = async (dataURI, name) => {
   });
   const file = new File([blob], "image.jpg");
   console.log("after: ", file);
+  const imgSrc = await UploadImage(file);
+  setImageWrap((prevArray) => {
+    const newArray = [...prevArray];
+    newArray[parseInt(name)] = imgSrc;
+    return newArray;
+  });
+  setLoading(false);
 };
