+++ src/pages/PostUpload.jsx
@@ -2,9 +2,10 @@ import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import { useRecoilValue } from "recoil";
 import { useNavigate } from "react-router-dom";
+
 //component
 import Layout from "../layout/Layout";
-import PostUploadWriting from "./../components/Post/PostUploadWriting";
+import TotalWritingUI from "../components/PostProductWriting/TotalWritingUI";
 
 //이미지
 import PostBackground from "../assets/post_bg.jpg";
@@ -16,26 +17,19 @@ import postingAPI from "../api/posting";
 //recoil
 import loginToken from "../recoil/loginToken";
 import accountname from "../recoil/accountname";
-import WritingUI from "../components/PostProductTest/WritingUI";
-
-// 작성중 다른곳으로 나가려고 할때 모달
-// 작성 완료시 업로드 할것인지 모달
-// 모든 작성 공간은 필수
 
 export default function PostUpload() {
   const navigate = useNavigate();
 
+  const token = useRecoilValue(loginToken);
   const account_name = useRecoilValue(accountname);
   const myProfile = `/profile/${account_name}`
 
   const [imageWrap, setImageWrap] = useState([]);
   const [userErrorMessage, setUserErrorMessage] = useState([]);
 
-  // 입력 데이터
+  // 게시글 입력 데이터
   const [postData, setPostData] = useState();
-
-  // 유저 토큰
-  const token = useRecoilValue(loginToken);
 
   const [data, setData] = useState({
     post: {
@@ -56,7 +50,6 @@ export default function PostUpload() {
 
   const handlePost = async (PostData, token) => {
     const response = await postingAPI(PostData, token);
-
     if (response.hasOwnProperty("post")) navigate(myProfile);
   };
 
@@ -75,25 +68,12 @@ export default function PostUpload() {
       errors.push("");
     }
     setUserErrorMessage(errors);
-
-    console.log("클릭");
   };
-
-  console.log(data);
 
   return (
     <Layout reduceTop="true">
       <PostProductWrap>
-        {/* <PostUploadWriting
-          textareaLabel="게시글 내용"
-          src={postUproad}
-          subtext="당신의 게시글을 업로드 해보세요!"
-          buttonText="게시글 업로드 하기"
-          showInput={false}
-          textareaHeight="300px"
-          getPostData={getPostData}
-        /> */}
-        <WritingUI
+        <TotalWritingUI
           src={postUproad}
           subtext="당신의 게시글을 업로드 해보세요!"
           getData={getPostData}
