+++ src/pages/ProductUpload.jsx
@@ -5,34 +5,28 @@ import { useNavigate } from "react-router-dom";
 
 //component
 import Layout from "../layout/Layout";
-import ProductUploadWriting from "../components/Product/ProductUploadWriting";
+import TotalWritingUI from "../components/PostProductWriting/TotalWritingUI";
 
 //이미지
 import PostBackground from "../assets/post_bg.jpg";
-import Product_upload from "../assets/Prodcut_upload.svg";
+import productUpload from "../assets/Prodcut_upload.svg";
 
 //API
 import UploadProductAPI from "../api/UploadProductAPI";
 
 //recoil
 import loginToken from "../recoil/loginToken";
-import WritingUI from "../components/PostProductTest/WritingUI";
-
-// 작성중 다른곳으로 나가려고 할때 모달
-// 작성 완료시 업로드 할것인지 모달
-// 모든 작성 공간은 필수
 
 export default function ProductUpload() {
   const navigate = useNavigate();
+
+  const token = useRecoilValue(loginToken);
 
   const [imageWrap, setImageWrap] = useState([]);
   const [userErrorMessage, setUserErrorMessage] = useState([]);
 
   // 상품 입력 데이터
   const [productData, setProductData] = useState();
-
-  // 유저 토큰
-  const token = useRecoilValue(loginToken);
 
   const [data, setData] = useState({
     product: {
@@ -86,21 +80,11 @@ export default function ProductUpload() {
     setUserErrorMessage(errors);
   };
 
-  console.log(data.product);
-
   return (
     <Layout reduceTop="true">
       <PostProductWrap>
-        {/* <ProductUploadWriting
-          src={Product_upload}
-          subtext="당신의 굿즈상품을 업로드하여 판매해보세요!"
-          buttonText="상품 업로드 하기"
-          showInput={true}
-          textareaHeight="100px"
-          getPostProductData={getPostProductData}
-        /> */}
-        <WritingUI
-          src={Product_upload}
+        <TotalWritingUI
+          src={productUpload}
           subtext="당신의 상품을 업로드 해보세요!"
           getData={getProductData}
           data={data}
